# Backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HTCMC PCHSystem",
    "version": "0.10.0"
  },
  "paths": {
...
    "/auth/token": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "签发一次性登录 token（MCDR）",
        "operationId": "post_token_auth_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenIssueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenIssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "X-Service-Token": []
          }
        ]
      }
    },
    "/auth/exchange": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "一次性 token 兑换 JWT 对",
        "operationId": "post_exchange_auth_exchange_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenExchangeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "续签 access token",
        "operationId": "post_refresh_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenExchangeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/me": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "当前身份（账号 + 活跃 UUID）",
        "description": "返回当前账号 + 绑定 players + active_uuid（会话来源 UUID）。\n\nplayer-less 托管账号（ADMIN_* env）无 active_uuid → null、players=[]（issue #74）。",
        "operationId": "get_me_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/me/last_sheet": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "我最近访问的表",
        "operationId": "get_my_last_sheet_me_last_sheet_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LastSheetResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          },
          {
            "X-Service-Token": []
          },
          {
            "X-Player-UUID": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AccountBrief": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "is_temporary": {
            "type": "boolean",
            "title": "Is Temporary"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "is_temporary",
          "role"
        ],
        "title": "AccountBrief",
        "description": "Web 账号摘要（与前端 Frontend/src/api/identity.ts AccountBrief 对齐）。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LastSheetResponse": {
        "properties": {
          "sheet_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sheet Id"
          }
        },
        "type": "object",
        "required": [
          "sheet_id"
        ],
        "title": "LastSheetResponse"
      },
...
      "MeResponse": {
        "properties": {
          "account": {
            "$ref": "#/components/schemas/AccountBrief"
          },
          "players": {
            "items": {
              "$ref": "#/components/schemas/app__schemas__auth__PlayerBrief"
            },
            "type": "array",
            "title": "Players"
          },
          "active_uuid": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Uuid"
          }
        },
        "type": "object",
        "required": [
          "account",
          "players",
          "active_uuid"
        ],
        "title": "MeResponse",
        "description": "当前身份响应（升级为 account + players + active_uuid）。\n\nactive_uuid 可空：player-less 托管管理账号（ADMIN_* env）无会话来源\nUUID（issue #74）。"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest"
      },
...
      "TokenExchangeRequest": {
        "properties": {
          "token": {
            "type": "string",
            "format": "uuid",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "TokenExchangeRequest"
      },
      "TokenExchangeResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "Bearer"
          },
          "player": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/app__schemas__auth__PlayerBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "account": {
            "$ref": "#/components/schemas/AccountBrief"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "account"
        ],
        "title": "TokenExchangeResponse",
        "description": "exchange / login / register / claim-bind 共用 shape（前端 AuthResponse）。\n\nplayer 永久账号必有至少一个绑定 player（!!PCH login 即自动挂临时账号）；\nregister/claim 边界场景下若暂无 player 允许 None，由调用方保证语义。"
      },
      "TokenIssueRequest": {
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid",
            "title": "Uuid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "uuid",
          "name"
        ],
        "title": "TokenIssueRequest"
      },
      "TokenIssueResponse": {
        "properties": {
          "login_url": {
            "type": "string",
            "title": "Login Url"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "previous_tokens_revoked": {
            "type": "integer",
            "title": "Previous Tokens Revoked"
          },
          "frontend_online": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frontend Online"
          }
        },
        "type": "object",
        "required": [
          "login_url",
          "expires_in",
          "previous_tokens_revoked"
        ],
        "title": "TokenIssueResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "app__schemas__auth__PlayerBrief": {
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid",
            "title": "Uuid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "uuid",
          "name",
          "role"
        ],
        "title": "PlayerBrief"
      },
...
    },
    "securitySchemes": {
      "X-Service-Token": {
        "type": "apiKey",
        "description": "服务端组件令牌（MCDR / 服务端 mod / 服主脚本）",
        "in": "header",
        "name": "X-Service-Token"
      },
      "Authorization": {
        "type": "apiKey",
        "description": "Web JWT 通道，值形如 \"Bearer <jwt>\"",
        "in": "header",
        "name": "Authorization"
      },
      "X-Player-UUID": {
        "type": "apiKey",
        "description": "MCDR 代理通道玩家 UUID（与 X-Service-Token 成对）",
        "in": "header",
        "name": "X-Player-UUID"
      },
      "X-Source-Id": {
        "type": "apiKey",
        "description": "施工上报源标识（服务端 mod 白名单名）",
        "in": "header",
        "name": "X-Source-Id"
      }
    }
  }
}